Updated methods and icons in CustomApplicationContext and Counter classes
Updated several methods in the `CustomApplicationContext` and `Counter` classes. The `ToString` method in `Counter` now omits the timestamp. Simplified instantiation of `ContextMenuStrip` and `ToolStripMenuItem` in `CreateContextMenu` and added a new `statsItem` menu item. Changed the icon file in `InitializeContext` to `favicon.ico` and updated the tooltip text. Modified `UpdateTrayIconText` to show a balloon tip with stats instead of setting the tooltip text. In the `MWH.KeyPressCounter.csproj` file, replaced `Icon1.ico` with `favicon.ico`.